configure,rbldnsd.c: replace mallinfo() with mallinfo2()
The mallinfo() function from malloc.h is deprecated, and has been
replaced by mallinfo2(). Additionally, the "mallinfo" struct that it
returns has been replaced by a "mallinfo2" struct. The only difference
between the two is that the newer struct contains members of type
size_t rather than int, which proved to be too small (leading to
overflows).

The call to ssprintf() that prints this information has been updated
to use the "z" length modifier, which is C99, but which is probably
safe by now. Other C99 features are already being used by rbldnsd.